Create WellsandSources
The purpose of this function is to return the number and position of wells and sources. 
If we take an incidence matrix and we put it inside the function as an input, It will return as output the number of wells and sources and also two matrices. 
One relating to sinks and other relating to the sources.
For each matrix the first column it reppresent the node (source or sink) and the second, the corresponding arc.