Create pathforeachHA
This function returns a Tensor T where each element is the hyperarch and
the first column are the nodes from which the hyperarch starts,and the second
column is formed by the nodes were the hyperarch enter
C is a tensor where each element is an hyperarch. The first column of each element specifies 
if the hyperarch enters, exits or doesn't interact with each node. The
second column rapresents the weight of the hyperarch.
A is the incidence matrix
V is the vector of the weights of each hyperarch