Create pathfinderv3
This function, given in input T, which takes from pathsforEachHA, the vector of weights, the nodes from which to start and those of the end. 

It will store the path in a matrix called PERCORSI. Then it sets the nodes to be find and starts again if could not found anything.  Otherwise, it updates the current path with the current node and the weight of the Hyperarc. 

Then it will explore the hypergraph in a random way using the variable k. 

The path is complete when the last node is the one, we were looking for. 

We put a limit to the dimension of the path, and this let us to avoid path that occur more than one time in the same cycle. “if t<=a*2”, to avoid the presence of cycle “if t>3”. 
We avoid cycles by eliminating paths that have sequences of repeated nodes and not by limiting the length of the paths.
In addiction, a path with the same node and weight of the others is a copy so is useless, and in any case, we must initialize again our variable.